+++ razu/razuconfig.py
@@ -1,4 +1,4 @@
-from .config import Config
+from config import Config
 
 
 class RazuConfig(Config):
